+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## XXX - 2021-XX
-- Add validateNameExpressions method to Domain
+- Add validateNameExpressions, getDomainNamePreviews methods to Domain
 
 ## 1.7.0 - 2021-12-02
 - Package updates.